separate out schedule components
