+++ client-course-schedulizer/src/components/reuseables/Schedule/index.ts
@@ -1,1 +1,4 @@
 export * from "./Schedule";
+export * from "./ScheduleBase";
+export * from "./ScheduleHeader";
+export * from "./ScheduleLeftTimeAxis";
